search: clean up lucky generator tests
