+++ internal/search/lucky/testdata/TestNewGenerator/rule_application#01.golden
