+++ internal/search/lucky/testdata/TestNewGenerator/rule_application#02.golden
