+++ internal/search/lucky/testdata/TestNewGenerator/rule_application.golden
